Member Application Update 3
This is the last update before I completely changed the concept and layout for the library application menu. Since there is only a few days left before we have to get this project done, I had to made some big changes for this application to make things easier for me and Tam.